PLAN-1783: Updating planning area layer component
