+++ src/interface/src/app/treatments/planning-area-layer/planning-area-layer.component.html
