+++ src/interface/src/app/treatments/treatment-map/treatment-map.component.html
@@ -22,8 +22,10 @@
     *ngIf="showMapControls$ | async"
     [mapLibreMap]="mapLibreMap"></app-map-controls>
 
-  <app-treatment-area-layer [mapLibreMap]="mapLibreMap">
-  </app-treatment-area-layer>
+  <app-planning-area-layer
+    [mapLibreMap]="mapLibreMap"
+    [polygonGeometry]="geometry">
+  </app-planning-area-layer>
 
   <app-map-project-areas
     [mapLibreMap]="mapLibreMap"
